objpool: Fix to make percpu slot allocation more robust
Since gfp & GFP_ATOMIC == GFP_ATOMIC is true for GFP_KERNEL | GFP_HIGH,
it will use kmalloc if user specifies that combination. Here the reason
why combining the __vmalloc_node() and kmalloc_node() is that the vmalloc
does not support all GFP flag, especially GFP_ATOMIC. So we should check
if gfp & (GFP_ATOMIC | GFP_KERNEL) != GFP_ATOMIC for vmalloc first. This
ensures caller can sleep.
And for the robustness, even if vmalloc fails, it should retry with
kmalloc to allocate it.

Fixes: aff1871bfc81 ("objpool: fix choosing allocation for percpu slots")
Signed-off-by: Masami Hiramatsu (Google) <mhiramat@kernel.org>