hung_check: Provide hung check timeout budget
Add hung check timeout budget API for theoletically possible
deadline of the hung_task watchdog by locks (currently mutex
is supported.)

This is not for a timeout caused by a task waiting for IO to
complete, but for a timeout caused by a task waiting for a
lock that the io-waiting task hold. With "Panic (Reboot) On
Hung Tasks" is enabled, such hung_task timeout causes a panic
and reboot. Therefore, if this option is enabled, any IO
retries done holding a lock may eventually be futile.

To prevent this, that code can reduce the number of retries
or return -ETIMEDOUT before doing IO based on the time budget
(== time to theoletical deadline, calulated by the time when
any lock acquired).

Signed-off-by: Masami Hiramatsu (Google) <mhiramat@kernel.org>